requirements.txt
fastapi
uvicorn[standard]==0.24.0
pydantic
python-multipart==0.0.6
python-dotenv==1.0.0
langgraph
langchain
langchain-community
langchain-core
httpx==0.25.2
langchain-groq
copilotkit==0.1.60
ag-ui-langgraph==0.0.7